@@ -18,28 +18,30 @@
     "tslint": "tslint \"app/**/*.ts\""
   },
   "dependencies": {
-    "@angular/core": "2.0.0",
     "@angular/common": "2.0.0",
     "@angular/compiler": "2.0.0",
+    "@angular/core": "2.0.0",
     "@angular/http": "2.0.0",
     "@angular/platform-browser": "2.0.0",
     "@angular/platform-browser-dynamic": "2.0.0",
     "@angular/platform-server": "2.0.0",
     "@angular/router": "3.0.0",
+    "@types/lodash": "^4.14.37",
+    "lodash": "^4.16.4",
     "nativescript-angular": "0.6.0",
-    "rxjs": "5.0.0-beta.12",
+    "nativescript-barcodescanner": "^1.3.5",
     "reflect-metadata": "^0.1.8",
-    "tns-core-modules": "2.2.1",
-    "nativescript-barcodescanner": "^1.3.5"
+    "rxjs": "5.0.0-beta.12",
+    "tns-core-modules": "2.3.0"
   },
   "devDependencies": {
-    "zone.js": "^0.6.21",
     "babel-traverse": "6.8.0",
     "babel-types": "6.8.1",
     "babylon": "6.8.0",
     "lazy": "1.0.11",
     "nativescript-dev-typescript": "^0.3.2",
     "tslint": "^3.14.0",
-    "typescript": "^1.8.10"
+    "typescript": "2.0.3",
+    "zone.js": "^0.6.21"
   }
 }
