@@ -8,6 +8,9 @@ import { NativeScriptRouterModule } from "nativescript-angular/router";
 import { AppComponent } from "./app.component";
 import { routes, navigatableComponents } from "./app.routes";
 
+import { ConfigService } from "./config/config.service";
+import { CartService } from "./shared/cart.service";
+
 @NgModule({
     bootstrap: [AppComponent],
     imports: [
@@ -20,7 +23,8 @@ import { routes, navigatableComponents } from "./app.routes";
     declarations: [
         AppComponent,
         navigatableComponents
-    ]
+    ],
+    providers: [ConfigService, CartService]
 })
 export class AppModule {}
 
