@@ -1,8 +1,7 @@
+import { NgModule } from "@angular/core";
 import { NativeScriptRouterModule } from "nativescript-angular/router";
 import { CommonModule } from '@angular/common';
-import { NgModule } from "@angular/core";
 import { HomeComponent } from "./home.component";
-
 
 @NgModule({
     imports: [CommonModule],
