@@ -1,14 +1,12 @@
 export class Product {
+    barcode: string;
     name: string;
     price: number;
-    barcode: string;
 
     constructor(_name: string, _price: number , _barcode: string) {
+        this.barcode = _barcode;
         this.name = _name;
         this.price = _price;
-        this.barcode = _barcode;
     }
-
-
 }
 
