@@ -1,6 +1,7 @@
 export const PRODUCTS =
-{
-    "673419133999": {
+[
+    {
+        barcode: "673419133999",
         name: "קוטג",
         price: 5,
         details: {
@@ -9,7 +10,8 @@ export const PRODUCTS =
             ]
         }
     },
-    "673419133920": {
+    {
+        barcode: "673419133920",
         name: "שוקולד השחר",
         price: 3,
         details: {
@@ -18,7 +20,8 @@ export const PRODUCTS =
             ]
         }
     },
-    "673419133913": {
+    {
+        barcode: "673419133913",
         name: "קפה שחור",
         price: 5.9,
         details: {
@@ -27,14 +30,16 @@ export const PRODUCTS =
             ]
         }
     },
-    "673419134002": {
+    {
+        barcode: "673419134002",
         name: "אורז",
         price: 7,
         details: {
             sales: []
         }
     },
-    "67344116": {
+    {
+        barcode: "67344116",
         name: "שמן זית",
         price: 8,
         details: {
@@ -42,7 +47,8 @@ export const PRODUCTS =
             ]
         }
     },
-    "673419134057": {
+    {
+        barcode: "673419134057",
         name: "חלב",
         price: 4,
         details: {
@@ -50,4 +56,4 @@ export const PRODUCTS =
             ]
         }
     }
-};
+];
