@@ -1,17 +1,20 @@
 import { Injectable } from "@angular/core";
-import { ScanService } from "./scan.service";
+import { Utils } from "./utils";
 import { PRODUCTS } from "./mocks/products_mock";
+import * as _ from "lodash";
 
 @Injectable()
 export class ProductsService {
-    constructor(private scanService:ScanService) {
+    utils: Utils;
+
+    constructor() {
     }
 
     scanProduct() {
-        return this.scanService.scan();
+        return this.utils.scan();
     }
 
-    search( barcode:string ) {
-        return PRODUCTS[barcode];
+    search( barcode: string ) {
+        return _.find(PRODUCTS, { barcode: barcode});
     }
 }
